todo_app_backend/pyproject.toml
[tool.poetry]
name = "todo-app-backend"
version = "0.1.0"
description = "MIT"
authors = ["fecaridade <caridade500@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
